cohorts/2023/04-deployment/homework/requirements.txt
boto3==1.26.155
click==8.1.3
pandas==2.0.2
scikit-learn==1.0.2
python-dotenv==1.0.0
pyarrow==12.0.1
fsspec==2023.6.0
s3fs==0.4.2